# platform/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Learning Platform API
  description: |
    REST API for the Information Security Learning Platform.  
    Provides authentication, user management, course management, enrollment,
    course-content hierarchy (chapters → sections → modules → materials),
    and test/quiz functionality.
  version: 1.0.0
  contact:
    name: Platform Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

# ─── Security Schemes ────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `POST /auth/login`.  
        Pass as `Authorization: Bearer <token>`.

  # ─── Reusable Schemas ───────────────────────────────────────────────
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "Introduction to Cybersecurity"
        description:
          type: string
          example: "Learn the fundamentals of information security."
        created_by:
          type: integer
          format: int64
          example: 2
        is_published:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: "2026-04-01T12:00:00Z"
    CreateCourseRequest:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          example: "Network Security Basics"
        description:
          type: string
          example: "Covers firewalls, VPNs, and IDS."
    PaginatedCourses:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10

    # ── Chapters ──
    Chapter:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        course_id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "Chapter 1: Threats"
        order_index:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
...
    Section:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        chapter_id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "Section 1.1: Malware"
        order_index:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
...
    Module:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        section_id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "Module: Ransomware Overview"
        order_index:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
    CreateModuleRequest:
      type: object
      required: [section_id, title, order_index]
      properties:
        section_id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "Module: Ransomware Overview"
        order_index:
          type: integer
          example: 1
    Material:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        module_id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "Lecture Slides"
        type:
          type: string
          description: "Type of material (e.g. video, pdf, text, link)"
          example: "pdf"
        file_url:
          type: string
          nullable: true
          example: "https://storage.example.com/slides.pdf"
        external_url:
          type: string
          nullable: true
          example: "https://youtube.com/watch?v=abc"
        content:
          type: string
          nullable: true
          example: "Inline text content here..."
        order_index:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
    CreateMaterialRequest:
      type: object
      required: [module_id, title, type, order_index]
      properties:
        module_id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "Lecture Slides"
        type:
          type: string
          example: "pdf"
        file_url:
          type: string
          nullable: true
          example: "https://storage.example.com/slides.pdf"
        external_url:
          type: string
          nullable: true
          example: null
        content:
          type: string
          nullable: true
          example: null
        order_index:
          type: integer
          example: 1

    # ── Enrollments ──
...
    CourseStructure:
      type: array
      items:
        type: object
        properties:
          chapter:
            $ref: "#/components/schemas/Chapter"
          sections:
            type: array
            items:
              type: object
              properties:
                section:
                  $ref: "#/components/schemas/Section"
                modules:
                  type: array
                  items:
                    $ref: "#/components/schemas/Module"

    # ── Common ──
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "success"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "error description"

# ─── Tags ─────────────────────────────────────────────────────────────
...
paths:
...
  /courses:
    get:
      tags: [Courses]
      summary: List published courses (paginated)
      description: Returns published courses with pagination. No authentication required.
      operationId: getPublicCourses
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Items per page
      responses:
        "200":
          description: Paginated list of published courses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCourses"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /courses/{id}:
    get:
      tags: [Courses]
      summary: Get a published course by ID
      description: Returns a single published course. No authentication required.
      operationId: getPublicCourseByID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Course ID
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Course not found or not published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Courses (Protected) ────────────────────────────────────────────
  /courses/:
    post:
      tags: [Courses]
      summary: Create a new course
      description: Creates a new course (unpublished). **Admin or Teacher only.**
      operationId: createCourse
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
      responses:
        "200":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — not admin/teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /courses/{id}/delete:
    delete:
      tags: [Courses]
      summary: Delete a course (soft delete)
      description: Soft-deletes a course by setting `deleted_at`. **Admin only.**
      operationId: deleteCourse
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Course ID
      responses:
        "200":
          description: Course deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: "course deleted"
        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /courses/{id}/publish:
    put:
      tags: [Courses]
      summary: Publish a course
      description: Marks a course as published. **Admin or Teacher only.**
      operationId: publishCourse
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Course ID
      responses:
        "200":
          description: Course published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: "published"
        "403":
          description: Forbidden — not admin/teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /courses/admin/all:
    get:
      tags: [Courses]
      summary: List all courses (admin)
      description: Returns all courses including unpublished ones. **Admin only.**
      operationId: getAllCoursesForAdmin
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of all courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
        "403":
          description: Forbidden — not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Course Structure (Public) ──────────────────────────────────────
  /courses/courses/{course_id}/structure:
    get:
      tags: [Courses]
      summary: Get full course structure
      description: |
        Returns the hierarchical structure of a course:  
        **chapters → sections → modules**.  
        No authentication required.
      operationId: getCourseStructure
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Course ID
      responses:
        "200":
          description: Course structure tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseStructure"
        "500":
          description: Failed to get chapters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Enrollments ────────────────────────────────────────────────────
...
  /modules/:
    post:
      tags: [Modules]
      summary: Create a module
      description: Creates a new module within a section. **Admin or Teacher only.**
      operationId: createModule
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModuleRequest"
      responses:
        "200":
          description: Module created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — not admin/teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /modules/course/{section_id}:
    get:
      tags: [Modules]
      summary: Get modules by section
      description: Returns all modules for a given section, ordered by `order_index`.
      operationId: getModulesBySection
      security:
        - BearerAuth: []
      parameters:
        - name: section_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Section ID
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /modules/material:
    post:
      tags: [Modules]
      summary: Create a material
      description: Creates a new learning material within a module. **Admin or Teacher only.**
      operationId: createMaterial
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaterialRequest"
      responses:
        "200":
          description: Material created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — not admin/teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /modules/{module_id}/materials:
    get:
      tags: [Modules]
      summary: Get materials by module
      description: Returns all materials for a given module, ordered by `order_index`.
      operationId: getMaterialsByModule
      security:
        - BearerAuth: []
      parameters:
        - name: module_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Module ID
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Tests ──────────────────────────────────────────────────────────
...